Create the module base-android and migrate things from the core Issue 2
