+++ base-android/proguard-rules.pro
