+++ base-android/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.urban.noise.base.android" />
