+++ base-android/src/main/java/com/urban/noise/base/android/di/modules/ContextModule.kt
@@ -15,18 +15,18 @@
  * limitations under the License.
  */
 
-package com.urban.noise.library.core.annotations
+package com.urban.noise.base.android.di.modules
 
-/**
- * This annotation allows us to open some classes for mocking purposes while they are final in
- * release builds.
- */
-@Target(AnnotationTarget.ANNOTATION_CLASS)
-annotation class OpenClass
+import android.app.Application
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
 
-/**
- * Annotate a class with [OpenForTesting] if you want it to be extendable in debug builds.
- */
-@OpenClass
-@Target(AnnotationTarget.CLASS)
-annotation class OpenForTesting
+@Module
+class ContextModule(private val application: Application) {
+
+    @Singleton
+    @Provides
+    fun provideContext(): Context = application
+}
