+++ buildSrc/src/main/kotlin/commons/android-dynamic-feature.gradle.kts
@@ -97,7 +97,7 @@ junitJacoco {
 
 dependencies {
     implementation(project(BuildModules.APP))
-    implementation(project(BuildModules.CORE))
+    implementation(project(BuildModules.BASE_ANDROID))
 
     implementation(Dependencies.KOTLIN)
     implementation(Dependencies.APPCOMPAT)
