+++ features/splashscreen/src/main/java/com/urban/noise/feature/splashscreen/ui/SplashscreenFragment.kt
@@ -19,7 +19,7 @@ package com.urban.noise.feature.splashscreen.ui
 
 import com.urban.noise.feature.splashscreen.R
 import com.urban.noise.feature.splashscreen.databinding.FragmentSplashscreenBinding
-import com.urban.noise.library.core.ui.base.BaseFragment
+import com.urban.noise.base.android.ui.base.BaseFragment
 
 class SplashscreenFragment : BaseFragment<FragmentSplashscreenBinding, SplashscreenViewModel>(
     layoutId = R.layout.fragment_splashscreen
